pyproject.toml
[tool.poetry]
name = "discord-bots"
version = "0.1.0"
description = ""
authors = ["Piyush Mishra <piyush_mishra31@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.28.1"
discord = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
